// src/api/generated/schemas/getUsersByIdLenses200DataItemType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Elysia Documentation
 * Development documentation
 * OpenAPI spec version: 0.0.0
 */

export type GetUsersByIdLenses200DataItemType = typeof GetUsersByIdLenses200DataItemType[keyof typeof GetUsersByIdLenses200DataItemType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetUsersByIdLenses200DataItemType = {
  TRIAL_GIFT: 'TRIAL_GIFT',
  ADMIN_GIFT: 'ADMIN_GIFT',
  ADMIN_ADJUSTMENT: 'ADMIN_ADJUSTMENT',
  TOP_UP: 'TOP_UP',
  AI_CHARGE: 'AI_CHARGE',
  AI_REFUND: 'AI_REFUND',
} as const;